# docs/swagger.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mini E-Commerce API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Carrito": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Obtiene el carrito del usuario autenticado.",
        "description": "Si el usuario no tiene carrito, el sistema crea uno automáticamente.\r\nRequiere autenticación.",
        "responses": {
          "200": {
            "description": "Carrito del usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "El usuarioId del token no es válido.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Carrito/agregar-item": {
      "post": {
        "tags": [
          "Carrito"
        ],
        "summary": "Agrega un item al carrito del usuario autenticado.",
        "description": "Requiere autenticación.\r\nSi el item ya existe en el carrito, se incrementa la cantidad.",
        "requestBody": {
          "description": "ProductoId y cantidad.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item agregado correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida (cantidad inválida o productId inválido).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario o producto no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Producto inactivo o stock insuficiente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Carrito/actualizar-cantidad-item": {
      "patch": {
        "tags": [
          "Carrito"
        ],
        "summary": "Actualiza la cantidad de un item del carrito.",
        "description": "Requiere autenticación.\r\nSi el item no existe en el carrito se devuelve error.",
        "requestBody": {
          "description": "ProductoId y nueva cantidad.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarritoItemAgregarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cantidad actualizada correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida (cantidad inválida o productId inválido).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, producto o item no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Producto inactivo o stock insuficiente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Carrito/eliminar-item/{carritoItemId}": {
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Elimina un item del carrito del usuario.",
        "description": "Requiere autenticación.\r\nSolo permite eliminar items que pertenezcan al carrito del usuario autenticado.",
        "parameters": [
          {
            "name": "carritoItemId",
            "in": "path",
            "description": "Id del item del carrito.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item eliminado correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "carritoItemId inválido.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario o item no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "El item no pertenece al usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Carrito/vaciar": {
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Vacía el carrito del usuario autenticado.",
        "description": "Requiere autenticación.\r\nElimina todos los items del carrito.",
        "responses": {
          "200": {
            "description": "Carrito vaciado correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "El usuarioId del token no es válido.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario o carrito no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CarritoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarritoItemDto"
            },
            "nullable": true
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CarritoDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "$ref": "#/components/schemas/CarritoDto"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "CarritoItemAgregarDto": {
        "required": [
          "productId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CarritoItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cartId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quatity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese: Bearer {tu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}